+++ src/main/java/com/sigmundgranaas/forgero/core/material/material/simple/SimpleMaterialPOJO.java
@@ -66,6 +66,6 @@ public class SimpleMaterialPOJO {
         public int miningLevel;
         public float miningSpeedAddition;
         public float attackSpeedAddition;
-        public int attackDamageAddition;
+        public float attackDamageAddition;
     }
 }
