+++ src/main/java/com/sigmundgranaas/forgero/core/material/material/simple/SimpleSecondaryMaterial.java
@@ -7,7 +7,7 @@ public interface SimpleSecondaryMaterial extends SecondaryMaterial {
 
     float getMiningSpeedAddition();
 
-    int getAttackDamageAddition();
+    float getAttackDamageAddition();
 
     float getAttackSpeedAddition();
 }
